requirements.txt
torch
numpy
Pillow
scipy==1.1.0
opencv-python
torchvision
h5py
tensorflow==1.3.0
tqdm
termcolor
matplotlib
seaborn
pandas
nltk
decord
